refactor:初步建立统一PendingAtom运行时
