+++ src/hivememory/alice/runtime/cache.py
@@ -7,250 +7,21 @@ Kernel Runtime 缓存实现。
 - 会话级生命周期（无需 LRU 淘汰）
 - UPDATE 后缓存失效支持
 
+注：PendingAtom 运行时缓存已迁移至 ``pending_atom/`` 子包
+（``PendingAtomRuntime``），本模块仅保留 ``KoakumaAtomCache`` /
+``AgentProfileCache``。
+
 作者: HiveMemory Team
 版本: 1.0
 """
 
 import logging
-import re
 from collections import OrderedDict
 from typing import Dict, List, Optional
-from uuid import uuid4
 
-from hivememory.alice.runtime.models import PendingAtom, RuntimeScope
-from hivememory.alice.runtime.pending_atom_state import (
-    PendingAtomResolution,
-    PendingAtomSnapshot,
-    PendingAtomStatus,
-)
-from hivememory.core.models import AgentProfile, Identity, MemoryAtom, MemoryType
-from hivememory.engines.generation.models import PendingAtomSettlement, UpdateFocus, WriteFocus
+from hivememory.core.models import AgentProfile, MemoryAtom, MemoryType
 
 logger = logging.getLogger(__name__)
-
-
-def _slugify(text: str, max_len: int = 30) -> str:
-    """将文本转为 alias 友好的 slug 片段。"""
-    slug = text.lower().strip()
-    slug = re.sub(r"[^a-z0-9\s_]", "", slug)
-    slug = re.sub(r"\s+", "_", slug)
-    slug = re.sub(r"_+", "_", slug).strip("_")
-    return slug[:max_len].rstrip("_")
-
-
-class PendingAtomCache:
-    """
-    L0 运行时 pending atom 缓存。
-
-    作为 Alice runtime 共享基础设施，主帧和子帧共享同一实例。
-    子帧 WRITE 后主帧自动可见，无需 merge。
-    """
-
-    def __init__(self) -> None:
-        self._atoms: Dict[str, PendingAtom] = {}
-        self._intent_index: Dict[str, str] = {}
-        self._redirects: Dict[str, PendingAtomSettlement] = {}
-        self._canonical_index: Dict[str, List[str]] = {}
-        self._resolution: Dict[str, PendingAtomResolution] = {}
-
-    def register_write(
-        self,
-        content: str,
-        title: Optional[str],
-        reason: Optional[str],
-        identity: Identity,
-        runtime_scope: Optional[RuntimeScope] = None,
-    ) -> PendingAtom:
-        """注册 WRITE pending atom，返回带有生成 alias 的 PendingAtom。"""
-        slug_source = title if title else content[:20]
-        slug = _slugify(slug_source)
-        if not slug:
-            slug = "untitled"
-        short_id = uuid4().hex[:4]
-        pending_alias = f"draft_{slug}_{short_id}"
-        intent_id = f"intent_{uuid4().hex[:12]}"
-        focus = WriteFocus(
-            content=content,
-            reason=reason,
-            title=title,
-            identity=identity,
-            pending_alias=pending_alias,
-            intent_id=intent_id,
-        )
-
-        atom = PendingAtom(
-            pending_alias=pending_alias,
-            intent_id=intent_id,
-            status=PendingAtomStatus.PENDING,
-            source_verb="WRITE",
-            focus=focus,
-            identity=identity,
-            runtime_scope=runtime_scope or RuntimeScope(),
-        )
-        self._atoms[pending_alias] = atom
-        self._intent_index[intent_id] = pending_alias
-        logger.debug(f"Registered pending WRITE: {pending_alias} (intent={intent_id})")
-        return atom
-
-    def register_update(
-        self,
-        base_alias: str,
-        base_uuid: str,
-        instruction: str,
-        content: Optional[str],
-        identity: Identity,
-        runtime_scope: Optional[RuntimeScope] = None,
-    ) -> PendingAtom:
-        """注册 UPDATE pending revision，返回带有生成 alias 的 PendingAtom。"""
-        short_id = uuid4().hex[:4]
-        pending_alias = f"rev_{base_alias}_{short_id}"
-        intent_id = f"intent_{uuid4().hex[:12]}"
-        focus = UpdateFocus(
-            instruction=instruction,
-            content=content or None,
-            base_alias=base_alias,
-            base_uuid=base_uuid,
-            identity=identity,
-            pending_alias=pending_alias,
-            intent_id=intent_id,
-        )
-
-        atom = PendingAtom(
-            pending_alias=pending_alias,
-            intent_id=intent_id,
-            status=PendingAtomStatus.PENDING,
-            source_verb="UPDATE",
-            focus=focus,
-            identity=identity,
-            runtime_scope=runtime_scope or RuntimeScope(),
-        )
-        self._atoms[pending_alias] = atom
-        self._intent_index[intent_id] = pending_alias
-        logger.debug(f"Registered pending UPDATE: {pending_alias} (intent={intent_id})")
-        return atom
-
-    def apply_settlement(self, settlement: PendingAtomSettlement) -> None:
-        """
-        Apply a settlement from the generation pipeline to update pending atom state.
-
-        Args:
-            settlement: PendingAtomSettlement instance
-        """
-        pending_alias = settlement.pending_alias
-        atom = self._atoms.get(pending_alias)
-
-        if atom is None and settlement.intent_id:
-            resolved_alias = self._intent_index.get(settlement.intent_id)
-            if resolved_alias:
-                atom = self._atoms.get(resolved_alias)
-
-        if atom is None:
-            logger.warning(
-                f"Settlement for unknown pending atom: "
-                f"alias={settlement.pending_alias}, intent={settlement.intent_id}"
-            )
-            return
-
-        atom.status = PendingAtomStatus.SETTLED
-        self._resolution[atom.pending_alias] = settlement.resolution
-
-        atom.settlement = settlement
-        if settlement.canonical_alias or settlement.canonical_uuid:
-            self._redirects[atom.pending_alias] = settlement
-            if settlement.canonical_uuid:
-                aliases = self._canonical_index.setdefault(
-                    settlement.canonical_uuid,
-                    [],
-                )
-                if atom.pending_alias not in aliases:
-                    aliases.append(atom.pending_alias)
-
-        logger.debug(
-            f"Settlement applied to '{atom.pending_alias}': "
-            f"resolution={settlement.resolution.value}, "
-            f"canonical={settlement.canonical_alias}"
-        )
-
-    def get(self, pending_alias: str) -> Optional[PendingAtom]:
-        """通过 pending alias 查询。"""
-        return self._atoms.get(pending_alias)
-
-    def get_by_intent_id(self, intent_id: str) -> Optional[PendingAtom]:
-        """通过 intent_id 查询 pending atom。"""
-        pending_alias = self._intent_index.get(intent_id)
-        if pending_alias:
-            return self._atoms.get(pending_alias)
-        return None
-
-    def get_redirect(self, pending_alias: str) -> Optional[PendingAtomSettlement]:
-        """返回 pending alias 对应的 canonical redirect settlement。"""
-        return self._redirects.get(pending_alias)
-
-    def get_pending_aliases_for_canonical_uuid(self, canonical_uuid: str) -> List[str]:
-        """返回指向同一 canonical UUID 的 pending alias 列表。"""
-        return list(self._canonical_index.get(canonical_uuid, []))
-
-    def has(self, alias: str) -> bool:
-        """检查 alias 是否为已注册的 pending atom。"""
-        return alias in self._atoms
-
-    def all_aliases(self) -> List[str]:
-        """返回所有已注册的 pending alias。"""
-        return list(self._atoms.keys())
-
-    def all_atoms(self) -> List[PendingAtom]:
-        """返回所有已注册的 PendingAtom。"""
-        return list(self._atoms.values())
-
-    def snapshot(self, pending_alias: str) -> Optional[PendingAtomSnapshot]:
-        """
-        返回 pending alias 对应的统一状态视图。
-
-        派生路径：
-        - 已结算（``self._resolution`` 命中）：status=SETTLED + 对应 resolution
-        - 未结算：直接读取 ``atom.status``（来自统一 ``PendingAtomStatus``）
-        """
-        atom = self._atoms.get(pending_alias)
-        if atom is None:
-            return None
-
-        resolution = self._resolution.get(pending_alias)
-        settlement = self._redirects.get(pending_alias)
-
-        if resolution is not None:
-            canonical_alias = settlement.canonical_alias if settlement else None
-            canonical_uuid = settlement.canonical_uuid if settlement else None
-            if not resolution.has_canonical:
-                canonical_alias = None
-                canonical_uuid = None
-            return PendingAtomSnapshot(
-                pending_alias=pending_alias,
-                status=PendingAtomStatus.SETTLED,
-                resolution=resolution,
-                canonical_alias=canonical_alias,
-                canonical_uuid=canonical_uuid,
-            )
-
-        return PendingAtomSnapshot(
-            pending_alias=pending_alias,
-            status=atom.status,
-            resolution=None,
-            canonical_alias=None,
-            canonical_uuid=None,
-        )
-
-    def clear(self) -> None:
-        """清空全部 pending atom。"""
-        self._atoms.clear()
-        self._intent_index.clear()
-        self._redirects.clear()
-        self._canonical_index.clear()
-        self._resolution.clear()
-
-    @property
-    def size(self) -> int:
-        """当前缓存的 pending atom 数量。"""
-        return len(self._atoms)
 
 
 class KoakumaAtomCache:
