+++ tests/unit/patchouli/mtp/test_read_chain.py
@@ -150,7 +150,7 @@ class TestReadAliasResolution:
         assert "readable" in result.response_content
 
     def test_read_redirected_pending_alias(self, koakuma):
-        pending = koakuma.pending_cache.register_write(
+        pending = koakuma.pending_runtime.register_write(
             content="pending content",
             title="Pending Note",
             reason=None,
@@ -161,7 +161,7 @@ class TestReadAliasResolution:
             alias="fact_canonical",
         )
         koakuma.atom_cache.ingest_atom(canonical)
-        koakuma.pending_cache.apply_settlement(
+        koakuma.pending_runtime.settle(
             PendingAtomSettlement(
                 pending_alias=pending.pending_alias,
                 intent_id=pending.intent_id,
@@ -186,13 +186,13 @@ class TestReadAliasResolution:
         ]
 
     def test_read_discarded_pending_alias(self, koakuma):
-        pending = koakuma.pending_cache.register_write(
+        pending = koakuma.pending_runtime.register_write(
             content="pending content",
             title="Pending Note",
             reason=None,
             identity=MTPExecutionContext().identity,
         )
-        koakuma.pending_cache.apply_settlement(
+        koakuma.pending_runtime.settle(
             PendingAtomSettlement(
                 pending_alias=pending.pending_alias,
                 intent_id=pending.intent_id,
